+++ android-exoplayer/src/main/java/com/brentvatne/exoplayer/ReactExoplayerView.java
@@ -11,6 +11,7 @@ import android.text.TextUtils;
 import android.util.Log;
 import android.view.View;
 import android.view.Window;
+import android.view.accessibility.CaptioningManager;
 import android.widget.FrameLayout;
 
 import com.brentvatne.react.R;
@@ -56,6 +57,7 @@ import com.google.android.exoplayer2.trackselection.MappingTrackSelector;
 import com.google.android.exoplayer2.trackselection.TrackSelection;
 import com.google.android.exoplayer2.trackselection.TrackSelectionArray;
 import com.google.android.exoplayer2.upstream.DataSource;
+import com.google.android.exoplayer2.upstream.DefaultAllocator;
 import com.google.android.exoplayer2.upstream.DefaultBandwidthMeter;
 import com.google.android.exoplayer2.upstream.cache.Cache;
 import com.google.android.exoplayer2.upstream.cache.CacheDataSourceFactory;
@@ -69,6 +71,7 @@ import java.lang.Math;
 import java.util.Map;
 import java.lang.Object;
 import java.util.ArrayList;
+import java.util.Locale;
 
 @SuppressLint("ViewConstructor")
 class ReactExoplayerView extends FrameLayout implements
@@ -97,7 +100,7 @@ class ReactExoplayerView extends FrameLayout implements
     private SimpleCache downloadCache;
     private DataSource.Factory mediaDataSourceFactory;
     private SimpleExoPlayer player;
-    private MappingTrackSelector trackSelector;
+    private DefaultTrackSelector trackSelector;
     private boolean playerNeedsSource;
 
     private int resumeWindow;
@@ -109,10 +112,18 @@ class ReactExoplayerView extends FrameLayout implements
     private boolean isBuffering;
     private float rate = 1f;
 
+    private int minBufferMs = DefaultLoadControl.DEFAULT_MIN_BUFFER_MS;
+    private int maxBufferMs = DefaultLoadControl.DEFAULT_MAX_BUFFER_MS;
+    private int bufferForPlaybackMs = DefaultLoadControl.DEFAULT_BUFFER_FOR_PLAYBACK_MS;
+    private int bufferForPlaybackAfterRebufferMs = DefaultLoadControl.DEFAULT_BUFFER_FOR_PLAYBACK_AFTER_REBUFFER_MS;
+
     // Props from React
     private Uri srcUri;
     private String extension;
     private boolean repeat;
+    private String audioTrackType;
+    private Dynamic audioTrackValue;
+    private ReadableArray audioTracks;
     private String textTrackType;
     private Dynamic textTrackValue;
     private ReadableArray textTracks;
@@ -233,7 +244,9 @@ class ReactExoplayerView extends FrameLayout implements
         if (player == null) {
             TrackSelection.Factory videoTrackSelectionFactory = new AdaptiveTrackSelection.Factory(BANDWIDTH_METER);
             trackSelector = new DefaultTrackSelector(videoTrackSelectionFactory);
-            player = ExoPlayerFactory.newSimpleInstance(getContext(), trackSelector, new DefaultLoadControl());
+            DefaultAllocator allocator = new DefaultAllocator(true, C.DEFAULT_BUFFER_SEGMENT_SIZE);
+            DefaultLoadControl defaultLoadControl = new DefaultLoadControl(allocator, minBufferMs, maxBufferMs, bufferForPlaybackMs, bufferForPlaybackAfterRebufferMs, -1, true);
+            player = ExoPlayerFactory.newSimpleInstance(getContext(), trackSelector, defaultLoadControl);
             player.addListener(this);
             player.setMetadataOutput(this);
             exoPlayerView.setPlayer(player);
@@ -319,7 +332,6 @@ class ReactExoplayerView extends FrameLayout implements
 
     private void releasePlayer() {
         if (player != null) {
-            isPaused = player.getPlayWhenReady();
             updateResumePosition();
             player.release();
             player.setMetadataOutput(null);
@@ -503,20 +515,43 @@ class ReactExoplayerView extends FrameLayout implements
     private void videoLoaded() {
         if (loadVideoStarted) {
             loadVideoStarted = false;
+            setSelectedAudioTrack(audioTrackType, audioTrackValue);
             setSelectedTextTrack(textTrackType, textTrackValue);
             Format videoFormat = player.getVideoFormat();
             int width = videoFormat != null ? videoFormat.width : 0;
             int height = videoFormat != null ? videoFormat.height : 0;
             eventEmitter.load(player.getDuration(), player.getCurrentPosition(), width, height,
-                    getTextTrackInfo());
+                    getAudioTrackInfo(), getTextTrackInfo());
         }
+    }
+
+    private WritableArray getAudioTrackInfo() {
+        WritableArray audioTracks = Arguments.createArray();
+
+        MappingTrackSelector.MappedTrackInfo info = trackSelector.getCurrentMappedTrackInfo();
+        int index = getTrackRendererIndex(C.TRACK_TYPE_AUDIO);
+        if (info == null || index == C.INDEX_UNSET) {
+            return audioTracks;
+        }
+
+        TrackGroupArray groups = info.getTrackGroups(index);
+        for (int i = 0; i < groups.length; ++i) {
+            Format format = groups.get(i).getFormat(0);
+            WritableMap audioTrack = Arguments.createMap();
+            audioTrack.putInt("index", i);
+            audioTrack.putString("title", format.id != null ? format.id : "");
+            audioTrack.putString("type", format.sampleMimeType);
+            audioTrack.putString("language", format.language != null ? format.language : "");
+            audioTracks.pushMap(audioTrack);
+        }
+        return audioTracks;
     }
 
     private WritableArray getTextTrackInfo() {
         WritableArray textTracks = Arguments.createArray();
 
         MappingTrackSelector.MappedTrackInfo info = trackSelector.getCurrentMappedTrackInfo();
-        int index = getTextTrackRendererIndex();
+        int index = getTrackRendererIndex(C.TRACK_TYPE_TEXT);
         if (info == null || index == C.INDEX_UNSET) {
             return textTracks;
         }
@@ -649,10 +684,10 @@ class ReactExoplayerView extends FrameLayout implements
         return false;
     }
 
-    public int getTextTrackRendererIndex() {
+    public int getTrackRendererIndex(int trackType) {
         int rendererCount = player.getRendererCount();
         for (int rendererIndex = 0; rendererIndex < rendererCount; rendererIndex++) {
-            if (player.getRendererType(rendererIndex) == C.TRACK_TYPE_TEXT) {
+            if (player.getRendererType(rendererIndex) == trackType) {
                 return rendererIndex;
             }
         }
@@ -726,12 +761,9 @@ class ReactExoplayerView extends FrameLayout implements
         this.repeat = repeat;
     }
 
-    public void setSelectedTextTrack(String type, Dynamic value) {
-        textTrackType = type;
-        textTrackValue = value;
-
-        int index = getTextTrackRendererIndex();
-        if (index == C.INDEX_UNSET) {
+    public void setSelectedTrack(int trackType, String type, Dynamic value) {
+        int rendererIndex = getTrackRendererIndex(trackType);
+        if (rendererIndex == C.INDEX_UNSET) {
             return;
         }
         MappingTrackSelector.MappedTrackInfo info = trackSelector.getCurrentMappedTrackInfo();
@@ -739,12 +771,20 @@ class ReactExoplayerView extends FrameLayout implements
             return;
         }
 
-        TrackGroupArray groups = info.getTrackGroups(index);
+        TrackGroupArray groups = info.getTrackGroups(rendererIndex);
         int trackIndex = C.INDEX_UNSET;
+
         if (TextUtils.isEmpty(type)) {
-            // Do nothing
-        } else if (type.equals("disabled")) {
-            trackSelector.setSelectionOverride(index, groups, null);
+            type = "default";
+        }
+
+        DefaultTrackSelector.Parameters disableParameters = trackSelector.getParameters()
+                .buildUpon()
+                .setRendererDisabled(rendererIndex, true)
+                .build();
+
+        if (type.equals("disabled")) {
+            trackSelector.setParameters(disableParameters);
             return;
         } else if (type.equals("language")) {
             for (int i = 0; i < groups.length; ++i) {
@@ -763,21 +803,65 @@ class ReactExoplayerView extends FrameLayout implements
                 }
             }
         } else if (type.equals("index")) {
-            trackIndex = value.asInt();
-        } else { // default. invalid type or "system"
-            trackSelector.clearSelectionOverrides(index);
-            return;
+            if (value.asInt() < groups.length) {
+                trackIndex = value.asInt();
+            }
+        } else { // default
+            if (rendererIndex == C.TRACK_TYPE_TEXT && Util.SDK_INT > 18 && groups.length > 0) {
+                // Use system settings if possible
+                CaptioningManager captioningManager
+                        = (CaptioningManager)themedReactContext.getSystemService(Context.CAPTIONING_SERVICE);
+                if (captioningManager != null && captioningManager.isEnabled()) {
+                    trackIndex = getTrackIndexForDefaultLocale(groups);
+                }
+            } else if (rendererIndex == C.TRACK_TYPE_AUDIO) {
+                trackIndex = getTrackIndexForDefaultLocale(groups);
+            }
         }
 
         if (trackIndex == C.INDEX_UNSET) {
-            trackSelector.clearSelectionOverrides(trackIndex);
+            trackSelector.setParameters(disableParameters);
             return;
         }
 
-        MappingTrackSelector.SelectionOverride override
-                = new MappingTrackSelector.SelectionOverride(
-                        new FixedTrackSelection.Factory(), trackIndex, 0);
-        trackSelector.setSelectionOverride(index, groups, override);
+        DefaultTrackSelector.Parameters selectionParameters = trackSelector.getParameters()
+                .buildUpon()
+                .setRendererDisabled(rendererIndex, false)
+                .setSelectionOverride(rendererIndex, groups,
+                        new DefaultTrackSelector.SelectionOverride(trackIndex, 0))
+                .build();
+        trackSelector.setParameters(selectionParameters);
+    }
+
+    private int getTrackIndexForDefaultLocale(TrackGroupArray groups) {
+        if (groups.length == 0) { // Avoid a crash if we try to select a non-existant group
+            return C.INDEX_UNSET;
+        }
+
+        int trackIndex = 0; // default if no match
+        String locale2 = Locale.getDefault().getLanguage(); // 2 letter code
+        String locale3 = Locale.getDefault().getISO3Language(); // 3 letter code
+        for (int i = 0; i < groups.length; ++i) {
+            Format format = groups.get(i).getFormat(0);
+            String language = format.language;
+            if (language != null && (language.equals(locale2) || language.equals(locale3))) {
+                trackIndex = i;
+                break;
+            }
+        }
+        return trackIndex;
+    }
+
+    public void setSelectedAudioTrack(String type, Dynamic value) {
+        audioTrackType = type;
+        audioTrackValue = value;
+        setSelectedTrack(C.TRACK_TYPE_AUDIO, audioTrackType, audioTrackValue);
+    }
+
+    public void setSelectedTextTrack(String type, Dynamic value) {
+        textTrackType = type;
+        textTrackValue = value;
+        setSelectedTrack(C.TRACK_TYPE_TEXT, textTrackType, textTrackValue);
     }
 
     public void setPausedModifier(boolean paused) {
@@ -864,5 +948,14 @@ class ReactExoplayerView extends FrameLayout implements
 
     public void setUseTextureView(boolean useTextureView) {
         exoPlayerView.setUseTextureView(useTextureView);
+    }
+
+    public void setBufferConfig(int newMinBufferMs, int newMaxBufferMs, int newBufferForPlaybackMs, int newBufferForPlaybackAfterRebufferMs) {
+        minBufferMs = newMinBufferMs;
+        maxBufferMs = newMaxBufferMs;
+        bufferForPlaybackMs = newBufferForPlaybackMs;
+        bufferForPlaybackAfterRebufferMs = newBufferForPlaybackAfterRebufferMs;
+        releasePlayer();
+        initializePlayer();
     }
 }