+++ android/src/main/java/com/brentvatne/react/ReactVideoView.java
@@ -1,14 +1,20 @@
 package com.brentvatne.react;
 
 import android.annotation.SuppressLint;
+import android.annotation.TargetApi;
+import android.app.Activity;
 import android.content.res.AssetFileDescriptor;
 import android.graphics.Matrix;
 import android.media.MediaPlayer;
+import android.media.TimedMetaData;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Handler;
 import android.util.Log;
 import android.view.MotionEvent;
+import android.view.WindowManager;
+import android.view.View;
+import android.view.Window;
 import android.webkit.CookieManager;
 import android.widget.MediaController;
 
@@ -18,6 +24,8 @@ import com.facebook.react.bridge.Arguments;
 import com.facebook.react.bridge.LifecycleEventListener;
 import com.facebook.react.bridge.ReadableMap;
 import com.facebook.react.bridge.WritableMap;
+import com.facebook.react.bridge.WritableArray;
+import com.facebook.react.bridge.WritableNativeArray;
 import com.facebook.react.uimanager.ThemedReactContext;
 import com.facebook.react.uimanager.events.RCTEventEmitter;
 import com.yqritc.scalablevideoview.ScalableType;
@@ -26,6 +34,7 @@ import com.yqritc.scalablevideoview.ScaleManager;
 import com.yqritc.scalablevideoview.Size;
 
 import java.io.IOException;
+import java.io.UnsupportedEncodingException;
 import java.util.HashMap;
 import java.util.Map;
 import java.lang.Math;
@@ -34,19 +43,30 @@ import java.math.BigDecimal;
 import javax.annotation.Nullable;
 
 @SuppressLint("ViewConstructor")
-public class ReactVideoView extends ScalableVideoView implements MediaPlayer.OnPreparedListener, MediaPlayer
-        .OnErrorListener, MediaPlayer.OnBufferingUpdateListener, MediaPlayer.OnCompletionListener, MediaPlayer.OnInfoListener, LifecycleEventListener, MediaController.MediaPlayerControl {
+public class ReactVideoView extends ScalableVideoView implements
+    MediaPlayer.OnPreparedListener,
+    MediaPlayer.OnErrorListener,
+    MediaPlayer.OnBufferingUpdateListener,
+    MediaPlayer.OnCompletionListener,
+    MediaPlayer.OnInfoListener,
+    LifecycleEventListener,
+    MediaController.MediaPlayerControl {
 
     public enum Events {
         EVENT_LOAD_START("onVideoLoadStart"),
         EVENT_LOAD("onVideoLoad"),
         EVENT_ERROR("onVideoError"),
         EVENT_PROGRESS("onVideoProgress"),
+        EVENT_TIMED_METADATA("onTimedMetadata"),
         EVENT_SEEK("onVideoSeek"),
         EVENT_END("onVideoEnd"),
         EVENT_STALLED("onPlaybackStalled"),
         EVENT_RESUME("onPlaybackResume"),
-        EVENT_READY_FOR_DISPLAY("onReadyForDisplay");
+        EVENT_READY_FOR_DISPLAY("onReadyForDisplay"),
+        EVENT_FULLSCREEN_WILL_PRESENT("onVideoFullscreenPlayerWillPresent"),
+        EVENT_FULLSCREEN_DID_PRESENT("onVideoFullscreenPlayerDidPresent"),
+        EVENT_FULLSCREEN_WILL_DISMISS("onVideoFullscreenPlayerWillDismiss"),
+        EVENT_FULLSCREEN_DID_DISMISS("onVideoFullscreenPlayerDidDismiss");
 
         private final String mName;
 
@@ -76,6 +96,10 @@ public class ReactVideoView extends ScalableVideoView implements MediaPlayer.OnP
     public static final String EVENT_PROP_WIDTH = "width";
     public static final String EVENT_PROP_HEIGHT = "height";
     public static final String EVENT_PROP_ORIENTATION = "orientation";
+    public static final String EVENT_PROP_METADATA = "metadata";
+    public static final String EVENT_PROP_TARGET = "target";
+    public static final String EVENT_PROP_METADATA_IDENTIFIER = "identifier";
+    public static final String EVENT_PROP_METADATA_VALUE = "value";
 
     public static final String EVENT_PROP_ERROR = "error";
     public static final String EVENT_PROP_WHAT = "what";
@@ -88,7 +112,6 @@ public class ReactVideoView extends ScalableVideoView implements MediaPlayer.OnP
     private Runnable mProgressUpdateRunnable = null;
     private Handler videoControlHandler = new Handler();
     private MediaController mediaController;
-
 
     private String mSrcUriString = null;
     private String mSrcType = "mp4";
@@ -106,6 +129,7 @@ public class ReactVideoView extends ScalableVideoView implements MediaPlayer.OnP
     private float mActiveRate = 1.0f;
     private boolean mPlayInBackground = false;
     private boolean mBackgroundPaused = false;
+    private boolean mIsFullscreen = false;
 
     private int mMainVer = 0;
     private int mPatchVer = 0;
@@ -191,6 +215,9 @@ public class ReactVideoView extends ScalableVideoView implements MediaPlayer.OnP
             mMediaPlayer.setOnBufferingUpdateListener(this);
             mMediaPlayer.setOnCompletionListener(this);
             mMediaPlayer.setOnInfoListener(this);
+            if (Build.VERSION.SDK_INT >= 23) {
+                mMediaPlayer.setOnTimedMetaDataAvailableListener(new TimedMetaDataAvailableListener());
+            }
         }
     }
 
@@ -207,6 +234,9 @@ public class ReactVideoView extends ScalableVideoView implements MediaPlayer.OnP
         if ( mMediaPlayer != null ) {
             mMediaPlayerValid = false;
             release();
+        }
+        if (mIsFullscreen) {
+            setFullscreen(false);
         }
     }
 
@@ -348,7 +378,6 @@ public class ReactVideoView extends ScalableVideoView implements MediaPlayer.OnP
     }
 
     public void setPausedModifier(final boolean paused) {
-
         mPaused = paused;
 
         if (!mMediaPlayerValid) {
@@ -371,6 +400,7 @@ public class ReactVideoView extends ScalableVideoView implements MediaPlayer.OnP
                 mProgressUpdateHandler.post(mProgressUpdateRunnable);
             }
         }
+        setKeepScreenOn(!mPaused);
     }
 
     // reduces the volume based on stereoPan
@@ -439,6 +469,39 @@ public class ReactVideoView extends ScalableVideoView implements MediaPlayer.OnP
         }
     }
 
+    public void setFullscreen(boolean isFullscreen) {
+        if (isFullscreen == mIsFullscreen) {
+            return; // Avoid generating events when nothing is changing
+        }
+        mIsFullscreen = isFullscreen;
+
+        Activity activity = mThemedReactContext.getCurrentActivity();
+        if (activity == null) {
+            return;
+        }
+        Window window = activity.getWindow();
+        View decorView = window.getDecorView();
+        int uiOptions;
+        if (mIsFullscreen) {
+            if (Build.VERSION.SDK_INT >= 19) { // 4.4+
+                uiOptions = SYSTEM_UI_FLAG_HIDE_NAVIGATION
+                        | SYSTEM_UI_FLAG_IMMERSIVE_STICKY
+                        | SYSTEM_UI_FLAG_FULLSCREEN;
+            } else {
+                uiOptions = SYSTEM_UI_FLAG_HIDE_NAVIGATION
+                        | SYSTEM_UI_FLAG_FULLSCREEN;
+            }
+            mEventEmitter.receiveEvent(getId(), Events.EVENT_FULLSCREEN_WILL_PRESENT.toString(), null);
+            decorView.setSystemUiVisibility(uiOptions);
+            mEventEmitter.receiveEvent(getId(), Events.EVENT_FULLSCREEN_DID_PRESENT.toString(), null);
+        } else {
+            uiOptions = View.SYSTEM_UI_FLAG_VISIBLE;
+            mEventEmitter.receiveEvent(getId(), Events.EVENT_FULLSCREEN_WILL_DISMISS.toString(), null);
+            decorView.setSystemUiVisibility(uiOptions);
+            mEventEmitter.receiveEvent(getId(), Events.EVENT_FULLSCREEN_DID_DISMISS.toString(), null);
+        }
+    }
+
     public void applyModifiers() {
         setResizeModeModifier(mResizeMode);
         setRepeatModifier(mRepeat);
@@ -456,7 +519,6 @@ public class ReactVideoView extends ScalableVideoView implements MediaPlayer.OnP
     public void setControls(boolean controls) {
         this.mUseNativeControls = controls;
     }
-
 
     @Override
     public void onPrepared(MediaPlayer mp) {
@@ -501,6 +563,9 @@ public class ReactVideoView extends ScalableVideoView implements MediaPlayer.OnP
                 }
             });
         }
+
+        // Select track (so we can use it to listen to timed meta data updates)
+        mp.selectTrack(0);
     }
 
     @Override
@@ -535,6 +600,9 @@ public class ReactVideoView extends ScalableVideoView implements MediaPlayer.OnP
 
     @Override
     public void onBufferingUpdate(MediaPlayer mp, int percent) {
+        // Select track (so we can use it to listen to timed meta data updates)
+        mp.selectTrack(0);
+
         mVideoBufferedDuration = (int) Math.round((double) (mVideoDuration * percent) / 100.0);
     }
 
@@ -581,21 +649,51 @@ public class ReactVideoView extends ScalableVideoView implements MediaPlayer.OnP
 
     @Override
     public void onCompletion(MediaPlayer mp) {
-
         isCompleted = true;
         mEventEmitter.receiveEvent(getId(), Events.EVENT_END.toString(), null);
+        if (!mRepeat) {
+            setKeepScreenOn(false);
+        }
+    }
+        
+    // This is not fully tested and does not work for all forms of timed metadata
+    @TargetApi(23) // 6.0
+    public class TimedMetaDataAvailableListener
+            implements MediaPlayer.OnTimedMetaDataAvailableListener
+    {
+        public void onTimedMetaDataAvailable(MediaPlayer mp, TimedMetaData data) {
+            WritableMap event = Arguments.createMap();
+
+            try {
+                String rawMeta  = new String(data.getMetaData(), "UTF-8");
+                WritableMap id3 = Arguments.createMap();
+
+                id3.putString(EVENT_PROP_METADATA_VALUE, rawMeta.substring(rawMeta.lastIndexOf("\u0003") + 1));
+                id3.putString(EVENT_PROP_METADATA_IDENTIFIER, "id3/TDEN");
+
+                WritableArray metadata = new WritableNativeArray();
+
+                metadata.pushMap(id3);
+
+                event.putArray(EVENT_PROP_METADATA, metadata);
+                event.putDouble(EVENT_PROP_TARGET, getId());
+            } catch (UnsupportedEncodingException e) {
+                e.printStackTrace();
+            }
+
+            mEventEmitter.receiveEvent(getId(), Events.EVENT_TIMED_METADATA.toString(), event);
+        }
     }
 
     @Override
     protected void onDetachedFromWindow() {
-
         mMediaPlayerValid = false;
         super.onDetachedFromWindow();
+        setKeepScreenOn(false);
     }
 
     @Override
     protected void onAttachedToWindow() {
-
         super.onAttachedToWindow();
 
         if(mMainVer>0) {
@@ -604,7 +702,7 @@ public class ReactVideoView extends ScalableVideoView implements MediaPlayer.OnP
         else {
             setSrc(mSrcUriString, mSrcType, mSrcIsNetwork, mSrcIsAsset, mRequestHeaders);
         }
-
+        setKeepScreenOn(true);
     }
 
     @Override