+++ examples/video-caching/index.js
@@ -1,0 +1,4 @@
+import { AppRegistry, Platform } from "react-native";
+import App from "./App";
+
+AppRegistry.registerComponent("VideoCaching", () => App);